refactor(cli): extract BookSearchCommands from BookCommands
- Migrate book search logic (`searchForBook`, `searchByTitle`, `searchByAuthor`, `searchByIsbn`) to the new `BookSearchCommands` class.
- Remove redundant search methods and unused imports from `BookCommands`.
- Improve separation of concerns by isolating book search-related commands into a dedicated class.
- Rename `BookCommands` to `BookCirculationCommands` for better alignment with its responsibilities.
