+++ src/main/java/com/penrose/bibby/cli/command/book/BookPlacementCommands.java
@@ -1,10 +1,9 @@
-package com.penrose.bibby.cli.command.book.organize;
+package com.penrose.bibby.cli.command.book;
 
 import com.penrose.bibby.cli.prompt.application.CliPromptService;
 import com.penrose.bibby.cli.prompt.domain.PromptOptions;
 import com.penrose.bibby.library.cataloging.book.contracts.dtos.BookDTO;
 import com.penrose.bibby.library.cataloging.book.contracts.ports.inbound.BookFacade;
-import com.penrose.bibby.library.stacks.bookcase.contracts.ports.inbound.BookcaseFacade;
 import com.penrose.bibby.library.stacks.shelf.contracts.dtos.ShelfDTO;
 import com.penrose.bibby.library.stacks.shelf.contracts.ports.inbound.ShelfFacade;
 import org.springframework.shell.command.annotation.Command;
